// ler/src/modules/grpc/proto/ler.pb.ts
export interface CreateShipmentReq {
  sender: User | undefined;
  receiver: User | undefined;
  trackingNumber: string;
  description: string;
  grandTotal: number;
  shipmentType: string;
  serviceName: string;
  isConsignment?: boolean | undefined;
}

export interface User {
  name: string;
  mobileNumber: string;
  address: string;
  email: string;
  userType: string;
  city: string;
  latitude?: string | undefined;
  longitude?: string | undefined;
}