// ler/src/modules/vendor/enum/courier-case.enum.ts
export enum CourierCaseEnum {
  STANDARD_RIYADH_JEDDAH_FM = 'standardRiyadhJeddahFM',
  STANDARD_OTHER_CITIES_FM = 'standardOtherCitiesFM',
  CONSIGNMENT_RIYADH_JEDDAH_FM = 'consignmentRiyadhJeddahFM',
  CONSIGNMENT_OTHER_CITIES_FM = 'consignmentOtherCitiesFM',
  CONSIGNMENT_RIYADH_LM = 'consignmentRiyadhLM',
  CONSIGNMENT_OTHER_CITIES_LM = 'consignmentOtherCitiesLM',
  STANDARD_FBS_RIYADH_LM = 'standardFbsRiyadhLM',
  STANDARD_OTHER_CITIES_LM = 'standardOtherCitiesLM',
} 